separated components and add style.ts
Separated out the individual components in the navbar and start implementing an external style.ts file
