+++ frontend/src/components/header/DesktopNav.tsx
@@ -1,0 +1,50 @@
+import { Box, Stack, useColorModeValue } from "@chakra-ui/react";
+
+export interface NavItem {
+  label: string;
+  subLabel?: string;
+  children?: Array<NavItem>;
+  href?: string;
+}
+
+export const NAV_ITEMS: Array<NavItem> = [
+  {
+    label: "Explore",
+    href: "explore",
+  },
+  {
+    label: "Products",
+    href: "products",
+  },
+  {
+    label: "Mentors",
+    href: "mentors",
+  },
+];
+export const DesktopNav = () => {
+  const linkColor = useColorModeValue("gray.600", "gray.200");
+  const linkHoverColor = useColorModeValue("gray.800", "white");
+
+  return (
+    <Stack direction={"row"} spacing={4}>
+      {NAV_ITEMS.map((navItem) => (
+        <Box key={navItem.label}>
+          <Box
+            as="a"
+            p={2}
+            href={navItem.href ?? "#"}
+            fontSize={"sm"}
+            fontWeight={500}
+            color={linkColor}
+            _hover={{
+              textDecoration: "none",
+              color: linkHoverColor,
+            }}
+          >
+            {navItem.label}
+          </Box>
+        </Box>
+      ))}
+    </Stack>
+  );
+};
