+++ frontend/src/components/header/MobileNav.tsx
@@ -1,0 +1,45 @@
+import {
+  Box,
+  Text,
+  Stack,
+  useColorModeValue,
+  useDisclosure,
+} from "@chakra-ui/react";
+import { NAV_ITEMS, NavItem } from "./DesktopNav";
+
+const MobileNavItem = ({ label, children, href }: NavItem) => {
+  const { onToggle } = useDisclosure();
+
+  return (
+    <Stack spacing={4} onClick={children && onToggle}>
+      <Box
+        py={2}
+        as="a"
+        href={href ?? "#"}
+        justifyContent="space-between"
+        alignItems="center"
+        textAlign={"center"}
+        _hover={{
+          textDecoration: "none",
+        }}
+      >
+        <Text
+          fontWeight={600}
+          color={useColorModeValue("gray.600", "gray.200")}
+        >
+          {label}
+        </Text>
+      </Box>
+    </Stack>
+  );
+};
+
+export const MobileNav = () => {
+  return (
+    <Stack bg={useColorModeValue("white", "gray.800")} display={{ md: "none" }}>
+      {NAV_ITEMS.map((navItem) => (
+        <MobileNavItem key={navItem.label} {...navItem} />
+      ))}
+    </Stack>
+  );
+};
