+++ frontend/src/components/header/NavBar.tsx
@@ -1,17 +1,16 @@
-"use client";
-
 import {
   Box,
   Flex,
-  Text,
   IconButton,
-  Stack,
   Collapse,
   useColorModeValue,
   useDisclosure,
   AbsoluteCenter,
 } from "@chakra-ui/react";
 import { HamburgerIcon, CloseIcon } from "@chakra-ui/icons";
+import { MobileNav } from "./MobileNav";
+import { DesktopNav } from "./DesktopNav";
+import { headerStyle } from "./style";
 
 export default function NavBar() {
   const { isOpen, onToggle } = useDisclosure();
@@ -19,6 +18,7 @@ export default function NavBar() {
     <AbsoluteCenter>
       <Box>
         <Flex
+          // style={headerStyle}
           bg={useColorModeValue("inherit", "gray.800")}
           color={useColorModeValue("gray.600", "white")}
           minH={"60px"}
@@ -29,7 +29,7 @@ export default function NavBar() {
           borderColor={useColorModeValue("gray.200", "gray.900")}
           align={"center"}
         >
-          <Flex mt={"25"} display={{ base: "flex", md: "none" }}>
+          <Flex display={{ base: "flex", md: "none" }}>
             <IconButton
               onClick={onToggle}
               icon={
@@ -63,90 +63,3 @@ export default function NavBar() {
     </AbsoluteCenter>
   );
 }
-
-const DesktopNav = () => {
-  const linkColor = useColorModeValue("gray.600", "gray.200");
-  const linkHoverColor = useColorModeValue("gray.800", "white");
-
-  return (
-    <Stack direction={"row"} spacing={4}>
-      {NAV_ITEMS.map((navItem) => (
-        <Box key={navItem.label}>
-          <Box
-            as="a"
-            p={2}
-            href={navItem.href ?? "#"}
-            fontSize={"sm"}
-            fontWeight={500}
-            color={linkColor}
-            _hover={{
-              textDecoration: "none",
-              color: linkHoverColor,
-            }}
-          >
-            {navItem.label}
-          </Box>
-        </Box>
-      ))}
-    </Stack>
-  );
-};
-
-const MobileNav = () => {
-  return (
-    <Stack bg={useColorModeValue("white", "gray.800")} display={{ md: "none" }}>
-      {NAV_ITEMS.map((navItem) => (
-        <MobileNavItem key={navItem.label} {...navItem} />
-      ))}
-    </Stack>
-  );
-};
-
-const MobileNavItem = ({ label, children, href }: NavItem) => {
-  const { isOpen, onToggle } = useDisclosure();
-
-  return (
-    <Stack spacing={4} onClick={children && onToggle}>
-      <Box
-        py={2}
-        as="a"
-        href={href ?? "#"}
-        justifyContent="space-between"
-        alignItems="center"
-        textAlign={"center"}
-        _hover={{
-          textDecoration: "none",
-        }}
-      >
-        <Text
-          fontWeight={600}
-          color={useColorModeValue("gray.600", "gray.200")}
-        >
-          {label}
-        </Text>
-      </Box>
-    </Stack>
-  );
-};
-
-interface NavItem {
-  label: string;
-  subLabel?: string;
-  children?: Array<NavItem>;
-  href?: string;
-}
-
-const NAV_ITEMS: Array<NavItem> = [
-  {
-    label: "Explore",
-    href: "explore",
-  },
-  {
-    label: "Products",
-    href: "products",
-  },
-  {
-    label: "Mentors",
-    href: "mentors",
-  },
-];
