+++ frontend/src/components/header/style.ts
@@ -1,0 +1,15 @@
+export const headerStyle: any = {
+    justifyContent: "space-between",
+    alignItems:"center",
+    height:"60px",
+    outline:1,
+    position:"absolute",
+    top:0,
+    left:0,
+    width:"100dvw",
+    padding:"0 var(--chakra-space-8)"
+}
+
+export const navbarFlexStyle: any = {
+    background: "useColorModeValue('inherit', 'gray.800')"
+}
